pyproject.toml
[tool.poetry]
name = "fis-tool"
version = "0.3.0"
description = "FIS (File Interaction Script) 工具，用于生成、读取和应用项目文件结构描述，方便与 LLM 进行交互，实现高效的代码变更。"
authors = ["KroMiose <li_xiangff@163.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
inquirer = "^3.2.4"
argparse = "^1.4.0"
google-generativeai = "^0.6.0"
gitignorefile = "^1.1.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
fis-tool = 'src.fis_tool:main'
